+++ Reproductor de Audio y Video Extendido/ListaRaices.h
@@ -8,14 +8,14 @@ class ListaRaices : public DWL::DListaEx {
 						       ~ListaRaices(void);
 	void						CrearListaRaiz(DWL::DhWnd *nPadre, const int cX, const int cY, const int cAncho, const int cAlto, const int cID);
 	DWL::DListaEx_Item         *AgregarRaiz(const wchar_t *nPath, const BOOL nRepintar = TRUE);
-	void						Evento_PintarSubItem(HDC hDC, const size_t NumItem, const size_t NumSubItem, RECT *Espacio);
+	void						Evento_PintarSubItem(HDC hDC, const LONGLONG NumItem, const LONGLONG NumSubItem, RECT *Espacio);
 
 	void                        Evento_MousePresionado(DWL::DEventoMouse &DatosMouse);
 	void                        Evento_MouseSoltado(DWL::DEventoMouse &DatosMouse);
   protected:
 	DWL::DListaIconos_Icono	  *_IconoPapelera;
 	DWL::DListaIconos_Icono	  *_IconoPapeleraResaltado;
-	size_t                     _ItemRaizPresionado;
-	size_t                     _SubItemRaizPresionado;
+	LONGLONG                   _ItemRaizPresionado;
+	LONGLONG                   _SubItemRaizPresionado;
 };
